purge v1 of rekey
